Redo rollup config to match new structure.
