+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "State management for Knockout apps.",
   "main": "lib/index.js",
-  "module": "es/index.js",
+  "module": "src/index.js",
   "dependencies": {},
   "devDependencies": {
     "babel-core": "^6.24.1",
@@ -23,16 +23,20 @@
   },
   "scripts": {
     "test": "karma start",
-    "build:es": "rollup --config rollup/rollup.es.config.js",
     "build:cjs": "rollup --config rollup/rollup.cjs.config.js",
     "build:umd": "rollup --config rollup/rollup.umd.config.js",
     "build:umd:min": "rollup --config rollup/rollup.umd.min.config.js",
-    "build": "npm run build:es && npm run build:cjs && npm run build:umd && npm run build:umd:min"
+    "build": "npm run build:cjs && npm run build:umd && npm run build:umd:min"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Spreetail/knockout-store.git"
   },
+  "files": [
+      "dist",
+      "lib",
+      "src"
+  ],
   "keywords": [
     "knockout",
     "state",
