+++ rollup/rollup.cjs.config.js
@@ -1,7 +1,7 @@
 import babel from 'rollup-plugin-babel';
-import mergeBaseConfig from './rollup.merge-base-config.js';
 
 const cjsConfig = {
+    entry: 'src/index.js',
     format: 'cjs',
     plugins: [
         babel({
@@ -11,6 +11,5 @@ const cjsConfig = {
     dest: 'lib/index.js'
 };
 
-const finalConfig = mergeBaseConfig(cjsConfig);
 
-export default finalConfig;
+export default cjsConfig;
