--- rollup/rollup.merge-base-config.js
@@ -1,9 +1,0 @@
-const baseConfig = {
-    entry: './src/index.js'
-};
-
-function mergeBaseConfig(config) {
-    return Object.assign({}, baseConfig, config);
-}
-
-export default mergeBaseConfig;
