+++ rollup/rollup.umd.config.js
@@ -1,7 +1,7 @@
 import babel from 'rollup-plugin-babel';
-import mergeBaseConfig from './rollup.merge-base-config.js';
 
-const cjsConfig = {
+const umdConfig = {
+    entry: 'src/index.js',
     format: 'umd',
     moduleName: 'ko.store',
     plugins: [
@@ -12,6 +12,4 @@ const cjsConfig = {
     dest: 'dist/knockout-store.js'
 };
 
-const finalConfig = mergeBaseConfig(cjsConfig);
-
-export default finalConfig;
+export default umdConfig;
