+++ src/index.js
@@ -1,8 +1,31 @@
-import connect from './connect';
-import { getState, setState } from './store';
+const stateObservable = ko.observable({});
 
-export default {
-    connect,
+function setState(state) {
+    stateObservable(state);
+}
+
+function getState() {
+    return stateObservable;
+}
+
+function connect(mapStateToParams, mergeParams) {
+    if (typeof mapStateToParams !== 'function') {
+        mapStateToParams = (state) => {};
+    }
+    if (typeof mergeParams !== 'function') {
+        mergeParams = Object.assign.bind(null, {});
+    }
+    return function(ViewModel) {
+        return function (params) {
+            const stateParams = mapStateToParams(stateObservable(), params);
+            const mergedParams = mergeParams(params, stateParams);
+            return new ViewModel(mergedParams);
+        };
+    };
+}
+
+export {
+    setState,
     getState,
-    setState
+    connect
 };
