--- src/store.js
@@ -1,14 +1,0 @@
-const stateObservable = ko.observable({});
-
-function setState(state) {
-    stateObservable(state);
-}
-
-function getState() {
-    return stateObservable;
-}
-
-export {
-    setState,
-    getState
-};
